fix(int-overflow): cast to size_t before multiplying width*height in sector/texture_manager
Same root cause as the tilemap resize overflow (PR #12): an int*int
multiplication overflows before the result is assigned to size_t. On
pathologically large editor levels (>65535^2 tiles) the surface calc could
wrap and mis-size the editor bounding box. Defensive cast before multiply.

+++ src/supertux/sector.cpp
@@ -633,7 +633,7 @@ Sector::get_editor_size() const
   size_t max_surface = 0;
   Size size;
   for (const auto& tilemap : get_all_tilemaps()) {
-    size_t surface = tilemap->get_width() * tilemap->get_height();
+    size_t surface = static_cast<size_t>(tilemap->get_width()) * static_cast<size_t>(tilemap->get_height());
     if (surface > max_surface) {
       max_surface = surface;
       size = tilemap->get_size();
